requirements.txt
Pillow
instagrapi
pysocks